refactor: mover componentes de paginas para a pasta pages
