+++ src/app/features/admin/pages/student-registration/student-registration.component.scss
@@ -1,5 +1,6 @@
 :host {
   display: grid;
+  grid-template-rows: 150px 1fr;
   margin: 0 auto;
   // TODO: configurar uma largura máxima para o conteúdo no estilo global
   max-width: 1020px;
@@ -9,7 +10,7 @@
   height: 100dvh;
 }
 
-.btn {
+.return-button {
   all: unset;
   align-self: start;
   justify-self: start;
@@ -17,18 +18,10 @@
   margin-block-start: 35px;
 }
 
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  justify-content: center;
-  width: 100%;
-}
-
 .section {
   align-self: start;
-  width: 100%;
   max-width: 550px;
+  width: 100%;
 
   &__title {
     margin-block-end: 0;
@@ -37,18 +30,23 @@
   &__description {
     margin-block-end: 1rem;
   }
-}
 
-::ng-deep app-button {
-  text-align: center;
+  .form {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    justify-content: center;
+    width: 100%;
 
-  button {
-    margin-block-start: 10px;
-    max-width: 265px;
+    &__btn {
+      max-width: 265px;
+      margin: 15px auto;
+      width: 100%;
+    }
   }
 }
 
-// TODO: mudar estilos globalmente ou criar um componente customizado para notificacoes
+// TODO: mudar estilos globalmente ou criar um componente customizado para notificações
 ::ng-deep .custom-snackbar {
   border-radius: 12px;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
